pyproject.toml
[project]
name = "dal-obscura"
version = "0.1.0"
description = "Data access layer with masking, row filters, and Arrow Flight"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "dal-obscura" }]

dependencies = [
  "pyarrow>=15.0.0",
  "duckdb>=1.0.0",
  "psutil>=5.9.0",
  "pyiceberg[sql-sqlite]>=0.6.0",
  "pydantic>=2.12.0",
  "pyjwt>=2.8.0",
  "sqlglot>=30.2.1",
  "cryptography>=46.0.7",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
]

[project.optional-dependencies]
server = ["grpcio>=1.62.0"]

test = [
  "pytest>=8.0.0",
  "pytest-benchmark>=4.0.0",
  "pytest-cov>=4.1.0",
  "httpx>=0.27.0",
]

[project.scripts]
dal-obscura = "dal_obscura.data_plane.interfaces.cli:main"
dal-obscura-control-plane = "dal_obscura.control_plane.interfaces.cli:main"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-q --benchmark-skip"

[tool.uv]
dev-dependencies = [
  "pre-commit>=3.4.0",
  "pytest>=8.0.0",
  "pytest-benchmark>=4.0.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.4.8",
  "httpx>=0.27.0",
  "sqlalchemy>=2.0.0",
  "ty>=0.0.22",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
  "C90",  # mccabe complexity
  "B",  # flake8-bugbear
  "C4",  # flake8-comprehensions
  "E",  # pycodestyle errors
  "F",  # Pyflakes
  "I",  # isort
  "PIE",  # flake8-pie
  "PGH",  # pygrep-hooks
  "RET",  # flake8-return
  "RSE",  # flake8-raise
  "RUF",  # Ruff-specific rules
  "SIM",  # flake8-simplify
  "TID252",  # flake8-tidy-imports: ban relative imports
  "UP",  # pyupgrade
]

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.format]
line-ending = "lf"
